[bson] Add mongo_bson_new_from_static_data().
This method will create a new MongoBson structure but will not duplicate
the buffers content and instead use it directly. The caller is responsible
to make sure the buffer outlives the structure. The GDestroyNotify will be
called when the reference is freed.

If you try to copy a MongoBson with a static buffer, the contents of the
buffer will be copied and a new structure returned.