[source] Add custom MongoSource to avoid g_idle_source_new() calls.
This saves a bunch of idle sources being created by
g_simple_async_result_complete_in_idle(). It helps speed things up
by about 20% in my basic throughput tests.

In the future, it would be nice to attach once of these sources to
a target GMainContext. But as that is not available now with how we
are using GSimpleAsyncResult, this is not a regression. We can add that
feature in the future.