[client] Start on request abstraction for queuing while disconnected.
The goal for this is to eventually get rid of the connect_async()
and possibly disconnect_async() methods. This will allow us to move
to a lazy connection style where the request is simply deferred until
we have connected. This also helps in the Replica Set case where we
might be finding the next PRIMARY node.